# server/common/swagger/Api.yaml
...
info:
  description: This is rest api for Checksheets
  version: 1.0.0
host: localhost:3000
basePath: /api/v1
produces:
  - application/json
schemes:
  - http
  - https
securityDefinitions: {}
swagger: '2.0'
paths:
...
  '/checksheet/:id':
    delete:
      parameters:
        - name: id
          in: path
          description: id of the Checksheet in the route URL
          required: true
          type: string
      description: Delete a Checksheet
      tags:
        - Checksheet
      responses:
        '200':
          description: Successfully deleted comment
...